refactor(layout): centralize layout styles + tighten menu item heights
Extract scattered sx objects into a single layout/styles.ts:
- move activityItemSx out of constants.ts (now numeric-only)
- relocate NAV_ROW_SX, navRowTextSlotProps, Sidebar container/header/list
  scroll sx, Layout mobileContentSx + mobile nav bar sx
- add activityBottomItemSx to unify the 4 duplicated inline sx blocks for
  ActivityBar bottom buttons (feedback / language / theme / user)

Tighten menu heights for a more compact nav:
- ActivityBar activity item minHeight 64 -> 48
- Sidebar nav row minHeight 52 -> 40, py 1.25 -> 0.75, mx 1.5 -> 1

No behavior change — pure style consolidation.
